Able to read json data from files
